// srcApp/pages/profile/model/types.ts
export interface UserProfileFormData {
  name: string;
  password: string;
  icon: string;
  permissions: string;
  registrationSources: string;
  instagramUrl: string;
  twitterUrl: string;
  mainTextContent: string;
  countryName: string;
  countryImageUrl: string;
  countryDescription: string;
  yourPhotoUrl: string;
  aboutYourself: string;
}